requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.32.0
billiard==4.2.1
blinker==1.9.0
boto3==1.35.63
botocore==1.35.63
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.3
dateparser==1.2.0
Django==5.1.4
django-cognito-jwt==0.0.4
django-cors-headers==4.6.0
django-extensions==3.2.3
django-otp==1.5.4
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.4.2
MarkupSafe==3.0.2
moment==0.12.1
msal==1.32.0
msrest==0.7.1
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.48
pycparser==2.22
PyJWT==2.10.1
pymongo==4.11.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2024.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.3
termcolor==2.5.0
times==0.7
tqdm==4.67.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
